refactor: change offerPrice to be of type double, ref. #1083

+++ src/main/java/MITELOVERS/dto/request/PlaceBidRequestDTO.java
@@ -15,5 +15,5 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class PlaceBidRequestDTO {
 
-    private BigDecimal offerPrice;
+    private double offerPrice;
 }
